+++ src/errors/index.ts
@@ -1,0 +1,5 @@
+export { ErrorBase } from './errorBase'
+export { InvalidData } from './invalidData'
+export { DeviceVersionUnsupported } from './deviceUnsupported'
+export { DeviceStatusCodes, DeviceStatusError } from './deviceStatusError'
+export { InvalidDataReason } from './invalidDataReason'