+++ src/parseUtils.ts
@@ -1,4 +1,5 @@
-import type { FixlenHexString, HexString, Uint8_t, Uint16_t, Uint32_t, Uint64_str, ValidBIP32Path, VarlenAsciiString, _Uint64_num, _Uint64_bigint } from "./types/internal";
+import { InvalidData, InvalidDataReason } from "./errors";
+import type { _Uint64_bigint, _Uint64_num, FixlenHexString, HexString, Uint8_t, Uint16_t, Uint32_t, Uint64_str, ValidBIP32Path, VarlenAsciiString } from "./types/internal";
 
 export const MAX_UINT_64_STR = "18446744073709551615";
 
@@ -62,77 +63,76 @@ export const isUintStr = (data: unknown, constraints: { min?: string, max?: stri
         )
 }
 
-export function validate(cond: boolean, errMsg: string): asserts cond {
-    // TODO: Error
-    if (!cond) throw new Error(errMsg)
+export function validate(cond: boolean, errMsg: InvalidDataReason): asserts cond {
+    if (!cond) throw new InvalidData(errMsg)
 }
 
 
-export function parseAscii(str: unknown, err: string): VarlenAsciiString {
-    validate(isString(str), err);
+export function parseAscii(str: unknown, errMsg: InvalidDataReason): VarlenAsciiString {
+    validate(isString(str), errMsg);
     validate(
         str.split("").every((c) => c.charCodeAt(0) >= 32 && c.charCodeAt(0) <= 126),
-        err
+        errMsg
     );
     return str as VarlenAsciiString
 }
 
 
-export function parseHexString(str: unknown, err: string): HexString {
-    validate(isHexString(str), err)
+export function parseHexString(str: unknown, errMsg: InvalidDataReason): HexString {
+    validate(isHexString(str), errMsg)
     return str
 }
 
-export function parseHexStringOfLength<L extends number>(str: unknown, length: L, err: string): FixlenHexString<L> {
-    validate(isHexStringOfLength(str, length), err)
+export function parseHexStringOfLength<L extends number>(str: unknown, length: L, errMsg: InvalidDataReason): FixlenHexString<L> {
+    validate(isHexStringOfLength(str, length), errMsg)
     return str
 }
 
-export function parseUint64_str(val: unknown, constraints: { min?: string, max?: string }, err: string): Uint64_str {
+export function parseUint64_str(val: unknown, constraints: { min?: string, max?: string }, errMsg: InvalidDataReason): Uint64_str {
     switch (typeof val) {
         case 'string':
-            validate(isUint64str(val) && isUintStr(val, constraints), err)
+            validate(isUint64str(val) && isUintStr(val, constraints), errMsg)
             return val
         case 'number':
-            validate(isUint64Number(val) && isUintStr(val.toString(), constraints), err)
+            validate(isUint64Number(val) && isUintStr(val.toString(), constraints), errMsg)
             return val.toString() as Uint64_str
         case 'bigint':
-            validate(isUint64Bigint(val) && isUintStr(val.toString(), constraints), err)
+            validate(isUint64Bigint(val) && isUintStr(val.toString(), constraints), errMsg)
             return val.toString() as Uint64_str
         default:
-            validate(false, err)
+            validate(false, errMsg)
     }
 }
 
-export function parseUint32_t(value: unknown, err: string): Uint32_t {
-    validate(isUint32(value), err)
+export function parseUint32_t(value: unknown, errMsg: InvalidDataReason): Uint32_t {
+    validate(isUint32(value), errMsg)
     return value
 }
 
-export function parseUint16_t(value: unknown, err: string): Uint16_t {
-    validate(isUint16(value), err)
+export function parseUint16_t(value: unknown, errMsg: InvalidDataReason): Uint16_t {
+    validate(isUint16(value), errMsg)
     return value
 }
 
-export function parseUint8_t(value: number, err: string): Uint8_t {
-    validate(isUint8(value), err)
+export function parseUint8_t(value: number, errMsg: InvalidDataReason): Uint8_t {
+    validate(isUint8(value), errMsg)
     return value
 }
 
-export function parseBIP32Path(value: unknown, err: string): ValidBIP32Path {
-    validate(isValidPath(value), err)
+export function parseBIP32Path(value: unknown, errMsg: InvalidDataReason): ValidBIP32Path {
+    validate(isValidPath(value), errMsg)
     return value
 }
 
 
-export function parseIntFromStr(str: string, err: string): number {
-    validate(isString(str), err)
+export function parseIntFromStr(str: string, errMsg: InvalidDataReason): number {
+    validate(isString(str), errMsg)
     const i = parseInt(str);
     // Check that we parsed everything
-    validate("" + i === str, err);
+    validate("" + i === str, errMsg);
     // Could be invalid
-    validate(!isNaN(i), err);
+    validate(!isNaN(i), errMsg);
     // Could still be float
-    validate(isInteger(i), err);
+    validate(isInteger(i), errMsg);
     return i;
 }